🚨 [Tests]: figma-infoテストファイルを機能別に分割
- figma-info.test.tsを3つのファイルに分割:
  - figma-info.from-url.test.ts (8テスト)
  - figma-info.from-object.test.ts (3テスト)
  - figma-info.create.test.ts (3テスト)
- describeブロックを削除しフラットな構造に変更
- itをtestに統一
- テスト名にメソッド名を含めて明確化
